saving work

+++ utils/GeometryObject.cpp
@@ -3,7 +3,7 @@
 #include <cassert>
 
 GeometryObject::GeometryObject(std::vector<utils::Vertex> vertices, std::vector<utils::Color> colors,
-                               std::vector<unsigned> indices)
+                               std::vector<size_t> indices)
     : m_vertices(std::move(vertices)),
       m_colors(std::move(colors)),
       m_indices(std::move(indices))
@@ -21,7 +21,7 @@ const std::vector<utils::Color>& GeometryObject::GetColors() const
     return m_colors;
 }
 
-const std::vector<unsigned>& GeometryObject::GetIndices() const
+const std::vector<size_t>& GeometryObject::GetIndices() const
 {
     return m_indices;
 }
@@ -48,7 +48,7 @@ GeometryObject makeRainbowCube()
         {0.6f, 0.0f, 0.8f},
         {1.0f, 0.4f, 0.7f},
     };
-    std::vector<unsigned int> indices = {
+    std::vector<size_t> indices = {
         0, 1, 3, 0, 3, 2,
         5, 4, 6, 5, 6, 7,
         1, 5, 7, 1, 7, 3,
@@ -68,7 +68,7 @@ GeometryObject makePlatform(float size)
         {-size, size, 0.0f},
     };
     std::vector<utils::Color> colors(4, {0.5f, 0.5f, 0.5f});
-    std::vector<unsigned int> indices = {
+    std::vector<size_t> indices = {
         0, 1, 2,
         0, 2, 3,
     };
