+++ utils/GeometryObject.h
@@ -8,17 +8,17 @@ class GeometryObject
 {
 public:
     GeometryObject(std::vector<utils::Vertex> vertices, std::vector<utils::Color> colors,
-        std::vector<unsigned> indices);
+        std::vector<size_t> indices);
 
     [[nodiscard]] const std::vector<utils::Vertex>& GetVertices() const;
     [[nodiscard]] const std::vector<utils::Color>& GetColors() const;
-    [[nodiscard]] const std::vector<unsigned>& GetIndices() const;
+    [[nodiscard]] const std::vector<size_t>& GetIndices() const;
 
 private:
     // HW_ITEM 10
     const std::vector<utils::Vertex> m_vertices;
     std::vector<utils::Color> m_colors;
-    const std::vector<unsigned int> m_indices;
+    const std::vector<size_t> m_indices;
 };
 
 GeometryObject makeRainbowCube();
